# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Audio Files Hosting API Specs",
    "version": "1.0"
  },
  "paths": {
...
    "/api/users/{userId}": {
      "delete": {
        "summary": "Delete a user by ID",
        "description": "Deletes a user by their ID. Only admin users are allowed to perform this action.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The ID of the user to delete.",
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User 123 deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or user does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "User does not exist"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (e.g., session expired).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Session expired"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (e.g., user is not an admin).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Forbidden to delete user"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}